requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cohere==4.23
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
dnspython==2.4.2
fastapi==0.98.0
fastavro==1.8.2
filelock==3.12.2
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langchain==0.0.283
langsmith==0.0.33
loguru==0.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pinecone-client==2.2.2
portalocker==2.7.0
protobuf==4.24.2
psutil==5.9.5
pydantic==1.10.9
pydub==0.25.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pywin32==306
PyYAML==6.0
qdrant-client==1.5.0
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.120
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
